requirements.txt
Django==5.2
djangorestframework==3.16.0
pillow==11.1.0
python-dotenv==1.1.0
django-phonenumber-field==8.1.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
celery==5.5.1